internal: test, Consolidate config loader test helper.
Move the static ConfigLoader registration into internal/test/config
so every package that needs isolation from the host git config can
call a single shared helper instead of duplicating the go:linkname
and plugin.Register boilerplate.

Addresses review feedback from @AriehSchneier and @pjbgf on PR #2122.
