+++ common_test.go
@@ -14,15 +14,13 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/go-git/go-git/v6/config"
+	testconfig "github.com/go-git/go-git/v6/internal/test/config"
 	"github.com/go-git/go-git/v6/plumbing"
 	"github.com/go-git/go-git/v6/plumbing/cache"
 	"github.com/go-git/go-git/v6/plumbing/format/packfile"
 	"github.com/go-git/go-git/v6/plumbing/object"
 	"github.com/go-git/go-git/v6/storage/filesystem"
 	"github.com/go-git/go-git/v6/storage/memory"
-	"github.com/go-git/go-git/v6/x/plugin"
-	xconfig "github.com/go-git/go-git/v6/x/plugin/config"
 )
 
 type BaseSuite struct {
@@ -56,21 +54,7 @@ func (s *BaseSuite) SetupSuite() {
 // default test user data. Tests that need specific config values should
 // register their own ConfigSource.
 func registerTestConfigLoader() {
-	resetPluginEntry("config-loader")
-
-	err := plugin.Register(plugin.ConfigLoader(), func() plugin.ConfigSource {
-		return xconfig.NewStatic(defaultTestConfig(), defaultTestConfig())
-	})
-	if err != nil {
-		panic(fmt.Errorf("failed to register config storers: %v", err))
-	}
-}
-
-func defaultTestConfig() config.Config {
-	cfg := config.NewConfig()
-	cfg.User.Name = "Test User"
-	cfg.User.Email = "test@example.com"
-	return *cfg
+	testconfig.RegisterDefault()
 }
 
 func (s *BaseSuite) buildBasicRepository() {
